+++ src/Datasets/Unlabeled.php
@@ -2,7 +2,9 @@
 
 namespace Rubix\ML\Datasets;
 
+use Rubix\ML\Transformers\Stateful;
 use Rubix\ML\Other\Helpers\DataType;
+use Rubix\ML\Transformers\Transformer;
 use Rubix\ML\Kernels\Distance\Distance;
 use InvalidArgumentException;
 
@@ -80,6 +82,25 @@ class Unlabeled extends DataFrame implements Dataset
         }
 
         return self::quick($samples);
+    }
+
+    /**
+     * Apply a transformation to the dataset.
+     *
+     * @param \Rubix\ML\Transformers\Transformer $transformer
+     * @return self
+     */
+    public function apply(Transformer $transformer) : self
+    {
+        if ($transformer instanceof Stateful) {
+            if (!$transformer->fitted()) {
+                $transformer->fit($this);
+            }
+        }
+
+        $transformer->transform($this->samples);
+
+        return $this;
     }
 
     /**