+++ tests/Transformers/GaussianRandomProjectorTest.php
@@ -53,7 +53,8 @@ class GaussianRandomProjectorTest extends TestCase
     {
         $this->expectException(RuntimeException::class);
 
-        $this->generator->generate(1)
-            ->apply($this->transformer);
+        $samples = $this->generator->generate(1)->samples();
+
+        $this->transformer->transform($samples);
     }
 }