+++ tests/Transformers/MinMaxNormalizerTest.php
@@ -54,7 +54,8 @@ class MinMaxNormalizerTest extends TestCase
     {
         $this->expectException(RuntimeException::class);
 
-        $this->generator->generate(1)
-            ->apply($this->transformer);
+        $samples = $this->generator->generate(1)->samples();
+
+        $this->transformer->transform($samples);
     }
 }