+++ tests/Transformers/OneHotEncoderTest.php
@@ -54,6 +54,8 @@ class OneHotEncoderTest extends TestCase
     {
         $this->expectException(RuntimeException::class);
 
-        $this->dataset->apply($this->transformer);
+        $samples = $this->dataset->samples();
+
+        $this->transformer->transform($samples);
     }
 }