+++ tests/Transformers/RobustStandardizerTest.php
@@ -50,7 +50,8 @@ class RobustStandardizerTest extends TestCase
     {
         $this->expectException(RuntimeException::class);
 
-        $this->generator->generate(1)
-            ->apply($this->transformer);
+        $samples = $this->generator->generate(1)->samples();
+
+        $this->transformer->transform($samples);
     }
 }