+++ tests/Transformers/TfIdfTransformerTest.php
@@ -56,6 +56,8 @@ class TfIdfTransformerTest extends TestCase
     {
         $this->expectException(RuntimeException::class);
 
-        $this->dataset->apply($this->transformer);
+        $samples = $this->dataset->samples();
+
+        $this->transformer->transform($samples);
     }
 }