+++ tests/Transformers/VarianceThresholdFilterTest.php
@@ -49,7 +49,8 @@ class VarianceThresholdFilterTest extends TestCase
     {
         $this->expectException(RuntimeException::class);
 
-        $this->generator->generate(1)
-            ->apply($this->transformer);
+        $samples = $this->generator->generate(1)->samples();
+
+        $this->transformer->transform($samples);
     }
 }